Move fail_on_log_exception to tests/conftest.py
