+++ tests/components/mqtt/test_alarm_control_panel.py
@@ -2,8 +2,6 @@
 import copy
 import json
 
-import pytest
-
 from homeassistant.components import alarm_control_panel
 from homeassistant.const import (
     STATE_ALARM_ARMED_AWAY,
@@ -553,7 +551,7 @@ async def test_discovery_update_alarm(hass, mqtt_mock, caplog):
     )
 
 
-@pytest.mark.no_fail_on_log_exception
+# @pytest.mark.no_fail_on_log_exception
 async def test_discovery_broken(hass, mqtt_mock, caplog):
     """Test handling of bad discovery message."""
     data1 = '{ "name": "Beer" }'
